refactor: separate user list into a new component
