+++ src/pages/users/[userId]/index.tsx
@@ -1,14 +1,6 @@
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { trpc } from "@utils/trpc";
 import MainLayout from "@components/MainLayout";
-import useOnScreen from "@hooks/useOnScreen";
-import PostCard from "@components/PostCard";
 import Image from "next/future/image";
 import ShouldRender from "@components/ShouldRender";
 import { useRouter } from "next/router";
@@ -23,13 +15,15 @@ import getUserDisplayName from "@utils/getUserDisplayName";
 import Popover from "@components/Popover";
 import dynamic from "next/dynamic";
 import { MdDelete, MdEditNote, MdOutlineTextSnippet } from "react-icons/md";
-import EmptyMessage from "@components/EmptyMessage";
 import { User } from "@utils/types";
 import Skeleton from "@components/Skeleton";
 import UserLinkPreview from "@components/EditAccountModal/UserLink/UserLinkPreview";
 import Button from "@components/Button";
-import Comment from "@components/Comment";
 import clsx from "clsx";
+
+const UserPageList = dynamic(() => import("@components/UserPageList"), {
+  ssr: false,
+});
 
 const FollowersModal = dynamic(
   () => import("@components/Follows/FollowersModal"),
@@ -76,10 +70,6 @@ const UserPage: React.FC = () => {
   const redirectToTerms = (value: "privacy" | "conduct") => () =>
     router.push(`/terms/${value}`);
 
-  const bottomRef = useRef<HTMLDivElement>(null);
-  const reachedBottom = useOnScreen(bottomRef);
-  const loadingArray = Array.from<undefined>({ length: 4 });
-
   const { data: user, isLoading: loadingUser } = trpc.useQuery(
     [
       "users.single-user",
@@ -96,67 +86,6 @@ const UserPage: React.FC = () => {
   );
 
   const { date, toggleDateType } = useGetDate(user?.createdAt);
-
-  const {
-    data: userComments,
-    isLoading: loadingComments,
-    isFetchingNextPage: fetchingMoreComments,
-    hasNextPage: hasMoreComments,
-    fetchNextPage: fetchMoreComments,
-  } = trpc.useInfiniteQuery(
-    [
-      "comments.user-comments",
-      {
-        userId,
-        limit: 4,
-        filter: currentFilter,
-      },
-    ],
-    {
-      getNextPageParam: (lastPage) => lastPage.nextCursor,
-      ssr: false,
-      enabled: isCommentsTab,
-    }
-  );
-
-  const commentsToShow = useMemo(
-    () => userComments?.pages.flatMap((page) => page.comments),
-    [userComments]
-  );
-
-  const noCommentsToShow =
-    !loadingComments && !commentsToShow?.length && !hasMoreComments;
-
-  const {
-    data,
-    isLoading: loadingPosts,
-    fetchNextPage: fetchMorePosts,
-    isFetchingNextPage: isFetchingMorePosts,
-    hasNextPage: hasMorePosts,
-  } = trpc.useInfiniteQuery(
-    [
-      "posts.posts",
-      {
-        userId,
-        limit: 4,
-        filter: currentFilter,
-      },
-    ],
-    {
-      enabled: isPostsTab,
-      getNextPageParam: (lastPage) => lastPage.nextCursor,
-      // The initial focus of this page is the user, so the
-      // user's comments can load afterwards.
-      ssr: false,
-    }
-  );
-
-  const dataToShow = useMemo(
-    () => data?.pages.flatMap((page) => page.posts),
-    [data]
-  );
-
-  const noPostsToShow = !loadingPosts && !dataToShow?.length && !hasMorePosts;
 
   const isDeleteAccountModalOpen = useState(false);
   const [, setIsDeleteAccountModalOpen] = isDeleteAccountModalOpen;
@@ -176,7 +105,7 @@ const UserPage: React.FC = () => {
   );
 
   const {
-    mutate,
+    mutate: deleteUser,
     error: deleteError,
     isLoading: deleting,
   } = trpc.useMutation(["users.delete-user"], {
@@ -273,20 +202,9 @@ const UserPage: React.FC = () => {
 
   const onConfirm = useCallback(() => {
     if (!!session?.user?.id) {
-      mutate({ userId: session.user.id });
+      deleteUser({ userId: session.user.id });
     }
-  }, [mutate, session]);
-
-  useEffect(() => {
-    if (reachedBottom && hasMorePosts && isPostsTab) {
-      fetchMorePosts();
-    }
-
-    if (reachedBottom && hasMoreComments && isCommentsTab) {
-      fetchMoreComments();
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [reachedBottom]);
+  }, [deleteUser, session]);
 
   useEffect(() => {
     if (deleteError) toast.error("Error deleting your account.");
@@ -312,7 +230,7 @@ const UserPage: React.FC = () => {
             />
             <ShouldRender if={!userIsProfileOwner && session?.user?.id}>
               <Button
-                disabled={loadingPosts}
+                disabled={loadingUser}
                 variant="gradient"
                 onClick={handleClickFollowButton}
                 absolute
@@ -491,51 +409,8 @@ const UserPage: React.FC = () => {
             </div>
           </div>
 
-          <div className="flex w-full flex-col gap-10">
-            <ShouldRender if={isPostsTab}>
-              {(loadingPosts ? loadingArray : dataToShow)?.map((post, i) => (
-                <PostCard
-                  post={post}
-                  key={loadingPosts ? i : post?.id}
-                  loading={loadingPosts}
-                />
-              ))}
-
-              <ShouldRender if={isFetchingMorePosts}>
-                <PostCard loading />
-              </ShouldRender>
-
-              <ShouldRender if={noPostsToShow}>
-                <EmptyMessage message="Hmm. It seems that this user has not created any posts yet." />
-              </ShouldRender>
-            </ShouldRender>
-
-            <ShouldRender if={isCommentsTab}>
-              {(loadingComments ? loadingArray : commentsToShow)?.map(
-                (comment, i) => (
-                  <Comment
-                    comment={comment}
-                    key={loadingComments ? i : comment?.id}
-                    loading={loadingComments}
-                    hideReplies
-                    variant="outlined"
-                    linkToPost
-                  />
-                )
-              )}
-
-              <ShouldRender if={fetchingMoreComments}>
-                <Comment loading linkToPost variant="outlined" />
-              </ShouldRender>
-
-              <ShouldRender if={noCommentsToShow}>
-                <EmptyMessage message="Hmm. It seems that this user has not commented on any posts yet." />
-              </ShouldRender>
-            </ShouldRender>
-          </div>
+          <UserPageList currentFilter={currentFilter} currentTab={currentTab} />
         </section>
-
-        <div ref={bottomRef} />
       </MainLayout>
 
       <ConfirmationModal
